requirements.txt
boto3==1.42.46
python-dotenv==1.2.1